requirements.txt
requests>=2.31.0
gpiozero>=2.0
python-dotenv>=1.0.0
RPi.GPIO>=0.7.1